requirements.txt
aiida-core>=2.5.0
fastmcp>=0.1.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
requests>=2.31.0
